reuse vorbis comment writing logic, set vendor always to "Mediabunny"
