+++ src/flac/flac-demuxer.ts
@@ -232,7 +232,8 @@ export class FlacDemuxer extends Demuxer {
 						const data = readBytes(pictureBlock, dataLength);
 						this.metadataTags.images ??= [];
 						this.metadataTags.images.push({
-							data,
+							// Sometimes readBytes returns a Uint8Array, sometimes a Buffer
+							data: new Uint8Array(data),
 							mimeType: mediaType,
 							// https://www.rfc-editor.org/rfc/rfc9639.html#table13
 							kind:
